# backend/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Universal ISP Platform API
  version: 1.1.0
  description: >
    Tenant-scoped (`org_id` enforced server-side), RBAC-protected API.
    Auth: Bearer JWT. Correlation: X-Request-ID (sent + echoed).
    Lists: ?search=&status=&page=&per_page=&sort=&dir= -> {data,total,page,per_page}.
    Without a database, data routes answer 503 explicitly (never fake rows).
servers:
  - url: /api/v1
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth: {type: http, scheme: bearer, bearerFormat: JWT}
  parameters:
    Search: {name: search, in: query, schema: {type: string}}
    Status: {name: status, in: query, schema: {type: string}}
    Page: {name: page, in: query, schema: {type: integer, default: 1}}
    PerPage: {name: per_page, in: query, schema: {type: integer, default: 25, maximum: 100}}
    Sort: {name: sort, in: query, schema: {type: string}}
    Dir: {name: dir, in: query, schema: {type: string, enum: [asc, desc]}}
paths:
  /auth/login:
    post:
      summary: Login (PBKDF2 users table; demo only with ISP_DEMO_LOGIN=1)
      security: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username: {type: string}
                password: {type: string}
      responses:
        '200': {description: '{token,user,must_rotate_password}'}
        '401': {description: invalid credentials}
        '429': {description: login throttled}
  /auth/logout:
    post:
      summary: Revoke current token (jti denylist)
      responses:
        '200': {description: logged out}
        '401': {description: invalid token}
  /customers:
    get: {summary: List customers (tenant-scoped), parameters: [{$ref: '#/components/parameters/Search'}, {$ref: '#/components/parameters/Status'}, {$ref: '#/components/parameters/Page'}, {$ref: '#/components/parameters/PerPage'}, {$ref: '#/components/parameters/Sort'}, {$ref: '#/components/parameters/Dir'}], responses: {'200': {description: paged list}, '503': {description: database not configured}}}
  /subscriptions:
    get: {summary: List subscriptions, parameters: [{$ref: '#/components/parameters/Search'}, {$ref: '#/components/parameters/Status'}, {$ref: '#/components/parameters/Page'}, {$ref: '#/components/parameters/PerPage'}], responses: {'200': {description: paged list}}}
...
  /noc/summary:
    get: {summary: Live NOC aggregates from inventory, responses: {'200': {description: counts}, '503': {description: database not configured}}}
...
  /lab/test:
    post:
      summary: Connection Lab test + capability discovery (credentials request-scoped, never stored)
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [vendor, family, connection_type]
              properties:
                vendor: {type: string}
                family: {type: string}
                connection_type: {type: string, enum: [cli, rest, cloud_api, restconf, netconf, ssh, radius, snmp, webhook, generic_http]}
                config: {type: object, additionalProperties: {type: string}}
      responses:
        '200': {description: '{healthy,result,error?}'}
        '404': {description: unknown connector}
...